// src/app/models/ariziModels/ariziSefer.ts
export interface AriziSefer {
  Id: number;
  UetdsSeferReferansNo: number;
  AracPlaka: string;
  HareketTarihi: string;
  HareketSaati: string;
  SeferAciklama: string;
  AracTelefonu: string;
  FirmaSeferNo: string;
  SeferBitisTarihi: string;
  SeferBitisSaati: string;
  AktifPasif: string;
  AktifPasifAciklama: string;
}